Struct blight::common::obb::OBB2D

pub struct OBB2D { /* fields omitted */ }

2D Oriented Bounding Box

Methods

impl OBB2D

pub fn new(center: Vector2, local: [Vector2; 2], radius: Vector2) -> Self

pub fn from_aabb(aabb: &AABB) -> Self

pub fn from_points(points: &[Vector2]) -> Self

Fit an OBB onto given points At least 3 points are required Note that this function has a complexity of O(n2) To avoid surprises the points must be in "clock-wise order" e.g. given the points: (1, 0) (0, 1), (1, 1), (0, 0) a correct ordering would be [(0, 0), (0, 1), (1, 1), (1, 0)]

 use blight::common::vector2::Vector2;
 use blight::common::obb::OBB2D;

 let points = [
     Vector2::new(3.0, 1.0),
     Vector2::new(2.0, 2.0),
     Vector2::new(4.0, 4.0),
     Vector2::new(5.0, 3.0),
 ];
 let result = OBB2D::from_points(&points); // Correct
 let points = [
     Vector2::new(3.0, 1.0),
     Vector2::new(5.0, 3.0),
     Vector2::new(4.0, 4.0),
     Vector2::new(2.0, 2.0),
 ];
 let result = OBB2D::from_points(&points); // Incorrect

pub fn get_center(&self) -> &Vector2

pub fn get_local(&self) -> &[Vector2; 2]

Borrow the local coordinate space matrix

pub fn get_extents(&self) -> &Vector2

pub fn rotation_matrix(&self) -> Matrix22

Note that this method creates a new Matrix on each call Use get_local if you want to avoid copying

pub fn intersects_aabb(&self, other: &AABB) -> bool

Check if OBB2D intersects with an AABB

pub fn intersects(&self, other: &OBB2D) -> bool

Check if OBB2D intersects with another OBB2D