Struct blight::common::quadtree::Quadtree

pub struct Quadtree<T> where
    T: Spacial,  { /* fields omitted */ }

Methods

impl<T> Quadtree<T> where
T: Spacial,

pub fn new(boundary: AABB) -> Self

pub fn insert(&mut self, point: T) -> Result<(), QuadtreeError>

Insert a single element into the tree

pub fn can_insert(&self, point: &T) -> Result<(), QuadtreeError>

Check if point can be inserted into the tree

pub fn insert_many(&mut self, points: Vec<T>) -> Result<(), QuadtreeError>

Insert a vector of elements

pub fn query_range<'a>(&'a self, range: AABB) -> Vec<&'a T>

Queries a range, returns a Vec of references to points inside range

pub fn query_range_static<'a>(&'a self, range: &AABB, result: &mut Vec<&'a T>)

Queries a given range and inserts the matching elements into the result Might be useful if you want to avoid repeated memory allocation