Struct blight::common::vector2::Vector2

pub struct Vector2 {
    pub x: f32,
    pub y: f32,
}

Fields
x: f32 y: f32
Methods

impl Vector2

pub fn new(x: f32, y: f32) -> Vector2

pub fn from_array(values: [f32; 2]) -> Vector2

pub fn orthogonal(&self) -> Vector2

Get a vector orthogonal to self

pub fn length(&self) -> f32

pub fn dot(&self, other: &Vector2) -> f32

Calculate dot product

pub fn sub(&self, other: &Vector2) -> Vector2

Subtract a vector from self returning a new vector

pub fn sub_mut(&mut self, other: &Vector2)

Subtract a vector from self

pub fn add(&self, other: &Vector2) -> Vector2

Add a vector to self returning a new vector

pub fn add_mut(&mut self, other: &Vector2)

Add a vector to self

pub fn get(&self, i: usize) -> f32

pub fn get_mut<'a>(&'a mut self, i: usize) -> &'a mut f32

pub fn set(&mut self, x: f32, y: f32)